Create RedditorMixin
